analyze/requirements.txt
# Python dependencies for EcoScan analyze module

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Backend API Server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Google AI APIs
google-generativeai>=0.3.0
google-cloud-vision>=3.4.0
